+++ StringHash.h
@@ -1,4 +1,4 @@
-/* $Id: StringHash.h,v 1.3 2004/04/08 15:02:30 blytkerchan Exp $ */
+/* $Id: StringHash.h,v 1.4 2004/05/07 16:28:21 blytkerchan Exp $ */
 /* Jail: Just Another Interpreted Language
  * Copyright (c) 2003-2004, Ronald Landheer-Cieslak
  * All rights reserved
@@ -47,7 +47,7 @@
 class StringHash : public Hash {
 public:
 	StringHash();
-	StringHash(ulong);
+	StringHash(unsigned int);
 	~StringHash();
 	bool read(char*);
 	
@@ -55,7 +55,7 @@ protected:
 	bool _write(void);
 	bool empty_key(void *key);
 	int cmp_keys(const void *, const void *);
-	ulong hash(const void *);
+	unsigned int hash(const void *);
 }; // StringHash
